+++ execute-generators/src/main/kotlin/de/itemis/mps/gradle/generate/Main.kt
@@ -3,16 +3,12 @@ package de.itemis.mps.gradle.generate
 import com.xenomachina.argparser.ArgParser
 import com.xenomachina.argparser.SystemExitException
 import com.xenomachina.argparser.mainBody
-import configureLogging
-import de.itemis.mps.gradle.project.loader.executeWithProject
-import org.apache.log4j.Logger
 
 fun main(args: Array<String>) = mainBody("execute-generators") {
     val parsed = ArgParser(args).parseInto(::GenerateArgs)
     var result = GenerationResult.Error
 
-    val logger = Logger.getLogger("de.itemis.mps.gradle.generate")
-    configureLogging(parsed.logLevel)
+    logging.configure(parsed.logLevel)
 
     try {
         result = parsed.buildLoader()