+++ project-loader/src/main/kotlin/de/itemis/mps/gradle/project/loader/EnvironmentKind.kt
@@ -1,0 +1,5 @@
+package de.itemis.mps.gradle.project.loader
+
+public enum class EnvironmentKind {
+    MPS, IDEA
+}