+++ project-loader/src/test/kotlin/ArgParsingTest.kt
@@ -1,4 +1,5 @@
 import com.xenomachina.argparser.ArgParser
+import de.itemis.mps.gradle.logging.LogLevel
 import de.itemis.mps.gradle.project.loader.Args
 import org.junit.Assert.assertEquals
 import org.junit.Test